api_gateway/requirements.txt
dotenv~=0.9.9
python-dotenv~=1.1.0
httpx~=0.28.1
fastapi~=0.115.12
uvicorn~=0.34.2
grpcio~=1.71.0
grpcio-tools~=1.71.0
protobuf~=5.29.5
pytest~=8.3.5
psycopg2-binary~=2.9.10
python-multipart~=0.0.20
uuid~=1.30